requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
filelock==3.13.4
Flask==3.0.3
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
packaging==24.0
platformdirs==4.2.0
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.3
waitress==3.0.0
Werkzeug==3.0.3
wheel==0.43.0
WTForms==3.1.2
yarl==1.9.4
